refactor: modularize integration test utilities and add team application test suite
